% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data.R
\name{get_covid_data}
\alias{get_covid_data}
\title{Function to crawl the CoVid-19 data for Germany}
\usage{
get_covid_data(level = "county")
}
\arguments{
\item{level}{string to define the level of data, either "county" or "state". Default is "county", which requests the county level data, where "state" gives the state level data.}
}
\value{
a data.frame object
}
\description{
This function leverages the NPGEO Corona Data Hub to extract and prepare the geo data.
}
